firestore/pyproject.toml
[tool.poetry]
name = "firestore"
version = "0.1.0"
description = ""
authors = ["github-actions <jojoto8845@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
firebase-admin = "^6.5.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
